gtp_u_kernel: Support file:open() returning FDesc module=raw_file_io_list for NSfd
In my erlang OTP-26 environment, the file:open() functions returns the
following as FDesc:
"""
{file_descriptor,raw_file_io_list,
 {file_descriptor,
  prim_file,
  #{handle => #Ref<0.114031393.2719088661.121660>,
    owner => <0.526.0>,
    r_buffer => #Ref<0.114031393.2719088642.121370>,
    r_ahead_size => 0}}}
"""

In this case, the OS file descriptor has to be obtained from the inner
FDesc structure, in the handle field. Furthermore, the pointer needs to
be converted to its integer using the prim_file:get_handle() function,
and finally converted to an integer.